식자재 갯수 int -> double 변경

+++ app/src/main/java/com/mop2022/team10/Rest/Model/IngredientModel.java
@@ -13,6 +13,6 @@ public class IngredientModel {
     public String img;
 
     //사용자별 정보
-    public int count;
+    public double count;
     public LocalDate expirationDate;
 }
